+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 =========
 
+## Unreleased
+
+- Avoid calling invokes with dependencies on unknown resources
+
 ## 1.0.0
 
 ### Improvements